+++ src/ui/ui.ts
@@ -74,9 +74,9 @@ export async function pickMacroFile(
     const openFileButton = options?.hideOpenPerItem
       ? undefined
       : {
-        iconPath: new vscode.ThemeIcon('go-to-file'),
-        tooltip: 'Open File',
-      };
+          iconPath: new vscode.ThemeIcon('go-to-file'),
+          tooltip: 'Open File',
+        };
     const items: UriQuickPickItem[] = createMacroFileItems(openFileButton);
     if (!options?.hideOpen) {
       items.unshift(QuickPickOpenFile, QuickPickConfigureSourceDirectories, {